+++ app/src/main/java/com/team/hogspot/model/relations/UserGeoSpotCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["userId", "geoSpotId"])
 data class UserGeoSpotCrossRef(
-    val userId: Int,
-    val geoSpotId: Int
+    val userId: Long,
+    val geoSpotId: Long
 )
